+++ src/main/java/racingcar/domain/MoveRule.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 @FunctionalInterface
 public interface MoveRule {
